ADD: realizar una reservasin requerir metodo de pago o acomodacion como dto obligatorio

+++ backend/src/main/java/com/tours/infrastructure/entities/booking/Pay.java
@@ -17,7 +17,7 @@ public class Pay {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Float paymentAmount;
+    private Double paymentAmount;
     private LocalDateTime PaymentDate;
 
     @ManyToOne
